requirements.txt
aniso8601==9.0.1
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.7.0
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
